backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
groq
redis
python-dotenv
